symbolic-web-calculator/requirements.txt
asgiref==3.2.5
Django==3.0.4
django-rest-framework==0.1.0
djangorestframework==3.11.0
mpmath==1.1.0
pytz==2019.3
sqlparse==0.3.1
sympy==1.5.1
